kube-aggregator: split out openapi spec priority sorter
